run/requirements.txt
google-adk>=1.23.0
a2a-sdk
uvicorn
